removed some warnings and fixed an include error

+++ src/orxonox/objects/weaponsystem/Munition.cc
@@ -355,7 +355,7 @@ namespace orxonox
         if (needed_magazines <= 0 && !this->bStackMunition_)
             return false;
 
-        if (amount <= needed_magazines)
+        if (amount <= (unsigned int)needed_magazines)
         {
             // We need more magazines than we get, so just add them
             this->magazines_ += amount;
